+++ hsts-client/src/main/java/hsts/client/gui/AskBotController.java
@@ -110,7 +110,7 @@ public class AskBotController extends GUIScreen {
 
     @FXML
     private void onBack() {
-        switchTo("/fxml/MainMenu.fxml", true);
+        switchTo("/fxml/MainMenu.fxml");
     }
 
     /**