+++ hsts-client/src/main/java/hsts/client/gui/ExamApprovalScreenController.java
@@ -141,7 +141,7 @@ public class ExamApprovalScreenController extends GUIScreen {
 
     @FXML
     private void onBack() {
-        switchTo("/fxml/MainMenu.fxml", true);
+        switchTo("/fxml/MainMenu.fxml");
     }
 
     // -----------------------------------------------------------------