+++ hsts-client/src/main/java/hsts/client/gui/GradingScreenController.java
@@ -222,7 +222,7 @@ public class GradingScreenController extends GUIScreen {
 
     @FXML
     private void onBack() {
-        switchTo("/fxml/MainMenu.fxml", true);
+        switchTo("/fxml/MainMenu.fxml");
     }
 
     /**