+++ hsts-client/src/main/java/hsts/client/gui/PrincipalBrowseController.java
@@ -199,7 +199,7 @@ public class PrincipalBrowseController extends GUIScreen {
 
     @FXML
     private void onBack() {
-        switchTo("/fxml/MainMenu.fxml", true);
+        switchTo("/fxml/MainMenu.fxml");
     }
 
     /**