+++ hsts-client/src/main/java/hsts/client/gui/QuestionMgmtController.java
@@ -261,7 +261,7 @@ public class QuestionMgmtController extends GUIScreen {
 
     @FXML
     private void onBack() {
-        switchTo("/fxml/MainMenu.fxml", true);
+        switchTo("/fxml/MainMenu.fxml");
     }
 
     // -----------------------------------------------------------------