+++ hsts-client/src/main/java/hsts/client/gui/ReportsController.java
@@ -129,7 +129,7 @@ public class ReportsController extends GUIScreen {
 
     @FXML
     private void onBack() {
-        switchTo("/fxml/MainMenu.fxml", true);
+        switchTo("/fxml/MainMenu.fxml");
     }
 
     // -----------------------------------------------------------------