+++ hsts-client/src/main/java/hsts/client/gui/TeacherReportsController.java
@@ -166,7 +166,7 @@ public class TeacherReportsController extends GUIScreen {
 
     @FXML
     private void onBack() {
-        switchTo("/fxml/MainMenu.fxml", true);
+        switchTo("/fxml/MainMenu.fxml");
     }
 
     /**